refactor(packages): share ensureParentFolders helper
