+++ src/services/packageExportService.ts
@@ -17,6 +17,7 @@ import {
 import { log } from "../logger/logManager";
 import { encodeToBase64 } from "../utils/base64";
 import { deepClone } from "../utils/deepClone";
+import { ensureParentFolders } from "../utils/ensureParentFolders";
 
 export interface BuildPackageOptions {
 	choices: IChoice[];
@@ -208,23 +209,4 @@ export async function writePackageToVault(
 	await ensureParentFolders(app, normalizedPath);
 	const serialized = JSON.stringify(pkg, null, 2);
 	await app.vault.adapter.write(normalizedPath, serialized);
-}
-
-async function ensureParentFolders(app: App, filePath: string): Promise<void> {
-	const lastSlash = filePath.lastIndexOf("/");
-	if (lastSlash < 0) return;
-
-	const folderPath = filePath.slice(0, lastSlash);
-	if (!folderPath) return;
-
-	const segments = folderPath.split("/").filter(Boolean);
-	let current = "";
-
-	for (const segment of segments) {
-		current = current ? `${current}/${segment}` : segment;
-		const exists = await app.vault.adapter.exists(current);
-		if (!exists) {
-			await app.vault.createFolder(current);
-		}
-	}
 }
