+++ src/services/packageImportService.ts
@@ -21,6 +21,7 @@ import { CommandType } from "../types/macros/CommandType";
 import { log } from "../logger/logManager";
 import { decodeFromBase64 } from "../utils/base64";
 import { deepClone } from "../utils/deepClone";
+import { ensureParentFolders } from "../utils/ensureParentFolders";
 
 export interface LoadedQuickAddPackage {
 	pkg: QuickAddPackage;
@@ -560,23 +561,6 @@ function findMultiByPath(
 	}
 
 	return currentMulti;
-}
-
-async function ensureParentFolders(app: App, filePath: string): Promise<void> {
-	const lastSlash = filePath.lastIndexOf("/");
-	if (lastSlash < 0) return;
-	const folderPath = filePath.slice(0, lastSlash);
-	if (!folderPath) return;
-
-	const segments = folderPath.split("/").filter(Boolean);
-	let current = "";
-	for (const segment of segments) {
-		current = current ? `${current}/${segment}` : segment;
-		const exists = await app.vault.adapter.exists(current);
-		if (!exists) {
-			await app.vault.createFolder(current);
-		}
-	}
 }
 
 function applyAssetPathOverrides(
